Finished inventory controller, updated services some

+++ Tiplr.Data/InventoryItem.cs
@@ -21,7 +21,7 @@ namespace Tiplr.Data
         [ForeignKey(nameof(Product))]
         public int ProductId { get; set; }
         public virtual Product Product { get; set; }
-        public double OnHandCount { get; set; }
+        public decimal OnHandCount { get; set; }
         public DateTimeOffset LastModifiedDtTm { get; set; }
 
         public string LastModifiedById  { get; set; }
